subscription_service/requirements.txt
# Django core
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-environ==0.11.2

# Celery / task queue
celery[redis]==5.4.0
redis==5.2.1

# Database
psycopg2-binary==2.9.10

# Payments
stripe==12.0.0

# Server
gunicorn==23.0.0
